problem: process was not stopped when I ran a second :Shell command while in a previous :Shell buffer
I think because we are using an unnamed scratch buffer, it makes sense
to make sure no process is running in background. maybe I will remove
this in the future to allow processes to continue to run in the
background if I use named buffers. But for now, I think this is cleaner.